requirements.txt
jira==3.10.5
python-dotenv==1.1.1
pyinstaller==6.16.0
pandas==2.3.2